Allow user-defined threshold to define loci as polymorphic
With this update there are two new arguments, threshold_poly_MSL and threshold_poly_NML to define the minimum number of ocurrences of each state (methylated/unmethylated for MSL, band/no band for NML) needed to define a locus polymorphic. There were an unconsistency as the manual said that value used to be 2, but that was changed in very early versions of msap to 1.
Closes #16